+++ example/a.ts
@@ -1,22 +1,11 @@
 import { Elysia } from '../src'
 
-const a = new Elysia({ name: 'p' }).trace(async ({ set, request }) => {
-	await request
-	console.log('A')
-	// await request
-	console.log('B')
-})
-
-const b = new Elysia({ scoped: true }).get('/scoped', () => 'hi')
-
 const app = new Elysia()
-	.onRequest(async () => {
-
+	.header({
+		'X-Powered-By': 'Elysia'
 	})
-	.use(a)
-	.use(b)
-	.get('/', () => 'hi')
+	.get('/', 'hi')
+	.get('/id/:id', ({ params: { id } }) => 'hi')
 	.listen(3000)
 
-// console.log(app.event)
-// console.log(app.routes[0].composed?.toString())
+console.log(app.routes[0].composed?.toString())